kuairand-like-prediction/requirements.txt

pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
lightgbm>=3.3
torch>=1.10
pyyaml
joblib
matplotlib
tqdm
jupyter
